feat: make csrf auth implementation

+++ src/Http/Controllers/Auth/LogoutController.php
@@ -7,15 +7,25 @@ use Illuminate\Http\JsonResponse;
 use Creopse\Creopse\Http\Controllers\Controller;
 use Creopse\Creopse\Enums\ResponseStatusCode;
 use Creopse\Creopse\Events\Auth\UserLoggedOutEvent;
+use Creopse\Creopse\Traits\DetectsMobileRequest;
+use Illuminate\Support\Facades\Auth;
 
 class LogoutController extends Controller
 {
+    use DetectsMobileRequest;
+
     /**
      * Destroy an authenticated session.
      */
     public function __invoke(Request $request): JsonResponse
     {
-        $request->user()->currentAccessToken()->delete();
+        if ($this->isMobileRequest($request)) {
+            $request->user()->currentAccessToken()->delete();
+        } else {
+            Auth::guard('web')->logout();
+            $request->session()->invalidate();
+            $request->session()->regenerateToken();
+        }
 
         event(new UserLoggedOutEvent($request->user()->id));
